[SCP-1] Add ability for issuers to mint tokens
This commit adds the ability for token issuers to mint tokens on-the-fly, every token starts with a supply of 0, and every issuer has the ability to mint any amount of tokens up-to the max supply.

This commit also encompasses a basic first-pass of the SCP Account-Based chain state processor, this layer of SCP works similarly to Ethereum, in which the ledger is Account-based instead of UTXO-based, tokens and accounts now have their own respective global/accounting methods for easy access to state data.