[SCP] Implement SCP-2, the Staking Token (1/2)
SCP-2 tokens are a more advanced version of SCP-2, with the capability to stake on a per-block, proportional basis to every holder of the token on the network, these tokens have additional parameters such as:
- The staking inflation (in sats) per-block.
- The minimum age (in blocks) an account must wait before they can stake, after performing any transaction, which resets their staking age timer.

SCP-2 tokens are fully compatible with any SCP-1 integrated services, making integration a breeze platform-wide.

Commit one-of-two, as the full implementation (of stake reward calculations and the redeem/claim contract method) will be done in the next day or two.