[UI] [API] Integrate full Lightwallet
The SCP Lightwallet has been integrated fully, it supports all functionalities of the SCP wallet, such as the full dashboard display of Balances, Activity, Sending/Receiving, Staking and APR/ROI calculations.

SCP is now fully independent from SCC Core and able to be ran out-of-the-box with zero configuration, like magic!