[SCP] Fix auto-claim overminting
Fixes an overminting bug with the supply calculations in an edge-case during 'receives', in which the stake was redeemed before the received amount was credited, bypassing a supply check for the received amount.

The logic was swapped, since stake redeems will always check the supply during the process, while debit/credit only checks at the start of the transfer, now, at most, the stake will be discarded and the send will operate like a regular SCP-1 transfer.