[UI] Settings improvements & wallet exports
The dark-theme CSS has been improved for the settings page, as well as some minor UX improvements, DB improvements.

This also includes the addition of a 'Wallet Export' option which allows users to export their keys in an encrypted form, this copy of the wallet is encrypted by the user's password and thus it's security is as great as the password that protects it.
This feature is suitable for transferring SCP Wallet files between multiple devices without manually importing private keys, and due to the encryption, it's better kitted to handle online transportation without immediate risk.