[UI] Add SCC Activity + Performance Settings
Full SCC activity can now be seen in SCP Wallet! The activity widget is split into two lists: SCC and SCP, the list can be switched instantly.

A new setting has been added which increases UI performance on very large wallets, this reduces the amount of UI load by only processing the last 200 transactions.

This can optionally be disabled if a user wants to see more than 200 TXs, but is not recommended as it will impede client performance.