[API] Add 403s for non-fullnode indexed endpoints
This will encourage users of the explorer-like indexed APIs to setup their own Full-nodes instead of relying too much on a fallback server, this also makes it more obvious which endpoints work on Full-nodes and which are allowed to use a Fallback server for.

Fallbacks are mostly recommended for basic wallet functions, useful for quick plug'n'play financial integrations.