[API] Add ability to enable/disable API modules + Ports
Individual API modules can now be enabled/disabled via config file, this allows hosts to customize and enhance their setups, security, or just use less resources on their daemon.

In addition, the API port for SCP can now be customized, the default SCP port is still 3000 and this will be used automatically if a custom option is not presented.