[SCP] Implement Token State Indexing
Tokens are now indexed in the SCP state by their relative position of the time of their deployment, this allows token contracts (64-char TX-ID strings) to be referred to by a single positional integer.

This addition saves ~25% of space within contract calls, at the slight expense of a few bytes extra memory usage for SCP Full-nodes, which is hardly a downfall even with 10,000s of tokens.

This upgrade, once activated, will increase the TPS of SCP by 25%, to an estimated block throughput of ~62.25 TPS, from the prior ~49.84 TPS.

And the fees for SCP have dropped ~25% in accordance with the lower byte usage.