[Wallet] [API] Fix listaddresses unlock status
The listaddresses unlock checks were previously just checking if addresses had an encrypted copy, but did not care if there was a decrypted copy, now we simply check for decrypted privkeys, and if none is available, the wallet is locked!

This will also perform slightly better due to using less DB-wrapper workarounds, instead using direct pointers.