[UI] Fix full-node IS activity duplication
Full-node activity was being duplicated due to a faulty IS activity cache system, this was replaced by simply removing the cache update for a direct RPC call to pull & parse the latest mempool data in real-time.

Lightwallets were not affected by this bug, as they rely on the Lightwallet server for activity data, which did not duplicate anything.