[SCP-4] Add fixes, full APIs and Documentation
The entire API set for interacting with SCP-4 has been added, allowing complete utilization, creation of collections, NFTs, read/write of state data, as well as full documentation for all implemented APIs for the Wallet and Tokens modules.

The 'supply' is now referenced to as "mints", for coins that have a max mint set, they cannot mint more (in total history) than this amount, this includes burns, burning an NFT will not reduce the mint count, meaning the max mints is a true maximum.