[API] [Config] Add "allowedips" setting
This adds a security setting which acts as a whitelisting system for IPs, essentially a simple built-in firewall system.

Some example usecases:
A local exchange hot-wallet would leave the option as default, which is localhost (127.0.0.1) only.

A networked exchange wallet, where the exchange and wallet are on separate servers, would add the exchange server's IP
(or proxy IP) into the wallet's config, this means the API can only be accessed by the exchange's host and/or proxy.

An explorer and/or public API would use the "all" setting, which allows any IP to access the exposed API modules without prior whitelisting, including localhost.